migrate all tests to new non-sketchy context
