--- packages/widget/backup/playwright.config.ts
@@ -1,85 +1,0 @@
-import { defineConfig, devices } from '@sand4rt/experimental-ct-web'
-import { widgetConfig } from './vite.config'
-
-const IS_CI = !!process.env.CI
-const ALL_BROWSERS = !!process.env.ALL_BROWSERS
-
-// Locally we only run tests in Chromium
-const localBrowsers = [
-  {
-    name: 'chromium',
-    use: { ...devices['Desktop Chrome'] },
-  },
-]
-
-// In CI, we run tests in all supported browsers
-const all_browsers = [
-  {
-    name: 'chromium',
-    use: { ...devices['Desktop Chrome'] },
-  },
-
-  {
-    name: 'firefox',
-    use: { ...devices['Desktop Firefox'] },
-  },
-
-  {
-    name: 'webkit',
-    use: { ...devices['Desktop Safari'] },
-  },
-
-  /* Test against mobile viewports. */
-  // {
-  //   name: 'Mobile Chrome',
-  //   use: { ...devices['Pixel 5'] },
-  // },
-  // {
-  //   name: 'Mobile Safari',
-  //   use: { ...devices['iPhone 12'] },
-  // },
-
-  /* Test against branded browsers. */
-  // {
-  //   name: 'Microsoft Edge',
-  //   use: { ...devices['Desktop Edge'], channel: 'msedge' },
-  // },
-  // {
-  //   name: 'Google Chrome',
-  //   use: { ...devices['Desktop Chrome'], channel: 'chrome' },
-  // },
-]
-
-/**
- * See https://playwright.dev/docs/test-configuration.
- */
-export default defineConfig({
-  testDir: './test/integration',
-  /* The base directory, relative to the config file, for snapshot files created with toMatchSnapshot and toHaveScreenshot. */
-  snapshotDir: './__snapshots__',
-  /* Maximum time one test can run for. */
-  timeout: 30 * 1000,
-  /* Run tests in files in parallel */
-  fullyParallel: true,
-  /* Fail the build on CI if you accidentally left test.only in the source code. */
-  forbidOnly: IS_CI,
-  /* Retry on CI only */
-  retries: IS_CI ? 2 : 0,
-  /* Opt out of parallel tests on CI. */
-  workers: IS_CI ? 1 : undefined,
-  /* Reporter to use. See https://playwright.dev/docs/test-reporters */
-  reporter: IS_CI ? [['list'], ['html']] : 'list',
-  /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
-  use: {
-    /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
-    trace: 'on-first-retry',
-
-    /* Port to use for Playwright component endpoint. */
-    ctPort: 3100,
-
-    ctViteConfig: widgetConfig,
-  },
-
-  /* Configure projects for major browsers */
-  projects: IS_CI || ALL_BROWSERS ? all_browsers : localBrowsers,
-})
