--- packages/widget/backup/playwright/index.ts
@@ -1,2 +1,0 @@
-// Import styles, initialize component theme here.
-// import '../src/common.css';
